Ahora adoptar_animal hace lo que dice el TP. Además 'cancelar' ya no es un nombre válido para un animal

+++ Archivos_auxiliares/funciones_auxiliares.cpp
@@ -28,7 +28,7 @@ string string_a_mayuscula(string texto){
 
 
 int string_a_entero( string texto ){
-    long unsigned int largo = texto.length();
+    long unsigned int largo = (long unsigned int)texto.length();
     int entero = 0;
     
     if (largo == 0) entero = 0;
